[lldb-dap] Move requests into their own object/file
